TEL-9285 - Fixing Chilean Numbers that are valid
Summary:
DP-13742: Fix to Indonesia toll free numbers
TEL-9285 - Fixing Chilean Numbers that are valid DP-17092 - Google Reverted the change and postponed the fix to Oct. We need to support both formats in the iterim. I used the file from the google fix in the last branch

Test Plan: NA

Reviewers: nick

Differential Revision: https://phabricator.fspot.us/D10246